remove kjar-module-after.jar and kjar-module-before.jar. Created projects for them under drools-test-coverage-jars-with-invoker to be tested in test-suite KieContainerTest
