+++ drools-test-coverage/drools-test-coverage-jars/drools-test-coverage-jars-with-invoker/src/it/kie-project-simple/src/main/resources/META-INF/kmodule.xml
