+++ drools-test-coverage/drools-test-coverage-jars/only-jar-pojo-not-kjar-no-kmodule/pom.xml
