+++ drools-test-coverage/drools-test-coverage-jars/only-jar-pojo-not-kjar-no-kmodule/src/main/java/org/drools/compiler/integrationtests/only_jar_pojo_not_kjar_no_kmodule/MyPojo.java
