+++ drools-test-coverage/drools-test-coverage-jars/testEnum/pom.xml
