+++ drools-test-coverage/drools-test-coverage-jars/testEnum/src/main/java/org/drools/TestEnum.java
