+++ drools-test-coverage/pom.xml
@@ -39,8 +39,8 @@
 
   <modules>
     <module>standalone</module>
+    <module>drools-test-coverage-jars</module>
     <module>test-suite</module>
-    <module>test-compiler-integration-jars</module>
     <module>test-compiler-integration</module>
     <module>test-integration-nomvel</module>
     <module>test-integration-notms</module>
